[refactor]:Socket user domain 분리
