+++ socket/src/main/java/Backend/socket/domain/chat/domain/ChatContent.java
@@ -13,15 +13,15 @@ public class ChatContent {
     private String image;
     private LocalDateTime time;
 
-    public static ChatContent createChatContent(String userName, String content,Chat chat) {
-        ChatContent chatContent = ChatContent.builder()
-                .userName(userName)
-                .content(content)
-                .time(LocalDateTime.now())
-                .build();
-        chat.addChatContent(chatContent);
-        return chatContent;
-    }
+//    public static ChatContent createChatContent(String userName, String content,Chat chat) {
+//        ChatContent chatContent = ChatContent.builder()
+//                .userName(userName)
+//                .content(content)
+//                .time(LocalDateTime.now())
+//                .build();
+//        chat.addChatContent(chatContent);
+//        return chatContent;
+//    }
     public static ChatContent createChatContent(String userName, String content, Room room,String image) {
         ChatContent chatContent = ChatContent.builder()
                 .userName(userName)
