--- socket/src/main/java/Backend/socket/domain/chat/repository/ChatRepository.java
@@ -1,8 +1,0 @@
-package Backend.socket.domain.chat.repository;
-
-
-import Backend.socket.domain.chat.domain.Chat;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface ChatRepository extends MongoRepository<Chat, String> {
-}
