+++ socket/src/main/java/Backend/socket/domain/chat/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package Backend.socket.domain.chat.repository;
 
-import Backend.socket.domain.chat.domain.User;
+import Backend.socket.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
